refactor: Phase 5 [5.6] migrate tracking/ — move common/ml_history/ to tracking/, add deprecation shims (BPR)
