+++ pipelines/train/mtl.pipe.py
@@ -13,7 +13,7 @@ import logging
 from typing import List, Tuple
 
 from data.folds import FoldCreator, TaskType
-from common.ml_history import MLHistory, DatasetHistory
+from tracking import MLHistory, DatasetHistory
 from configs.globals import CATEGORIES_MAP
 from training.runners.mtl_cv import train_with_cross_validation
