+++ pipelines/train/next_head.pipe.py
@@ -13,7 +13,7 @@ from configs.paths import IoPaths, RESULTS_ROOT, EmbeddingEngine
 from configs.experiment import ExperimentConfig
 from data.folds import FoldCreator, TaskType
 from training.runners.next_cv import run_cv
-from common.ml_history import MLHistory, DatasetHistory
+from tracking import MLHistory, DatasetHistory
 from configs.globals import CATEGORIES_MAP
 
 logging.basicConfig(
