+++ src/tracking/experiment.py
@@ -1,0 +1,194 @@
+from __future__ import annotations
+
+import time
+from pathlib import Path
+from typing import Dict, List, Optional, Set, Union
+
+from tracking.fold import FoldHistory
+from tracking.parms.neural import HyperParams
+from tracking.utils.dataset import DatasetHistory
+from tracking.utils.time_history import TimeHistory
+
+
+class FlopsMetrics:
+    """Holds model profiling metrics."""
+
+    def __init__(self, flops, params):
+        self.flops = flops
+        self.params = params
+        self.memory: List[float] = []
+        self.inference_time: List[float] = []
+        self.training_time: List[float] = []
+
+    def to_dict(self) -> dict:
+        return {
+            'flops': self.flops,
+            'params': self.params,
+            'memory': self.memory,
+            'inference_time': self.inference_time,
+            'training_time': self.training_time,
+        }
+
+
+class MLHistory:
+    """Top-level experiment tracker for cross-validated ML training.
+
+    Manages folds, timing, model metadata, and delegates to
+    HistoryDisplay and HistoryStorage for output.
+
+    Usage:
+        history = MLHistory(
+            'MyModel',
+            tasks={'next', 'category'},
+            num_folds=5,
+            label_map={0: 'Food', 1: 'Shop'},
+            save_path='/results',
+            verbose=True,
+        )
+
+        with history:
+            for fold in folds:
+                # training loop...
+                history.step()
+        # end_training() auto-fires, results auto-saved
+    """
+
+    def __init__(
+        self,
+        model_name: str,
+        tasks: Set[str] | str,
+        model_type: str = 'MTL',
+        num_folds: int = 1,
+        model_parms: Optional[HyperParams] = None,
+        datasets: Optional[Set[DatasetHistory]] = None,
+        monitor: str = 'f1',
+        mode: str = 'max',
+        label_map: Optional[Dict[int, str]] = None,
+        save_path: Optional[Union[str, Path]] = None,
+        verbose: bool = False,
+        display_report: bool = False,
+    ):
+        self.model_name = model_name
+        self.model_type = model_type
+        self.num_folds = num_folds
+        self.model_parms: Optional[HyperParams] = model_parms
+        self.model_arch: Optional[str] = None
+        self.datasets: Optional[Set[DatasetHistory]] = datasets
+        self.monitor = monitor
+        self.mode = mode
+
+        self.tasks: Set[str] = {tasks} if isinstance(tasks, str) else tasks
+        self.folds: List[FoldHistory] = [
+            FoldHistory(i, self.tasks, monitor=monitor, mode=mode)
+            for i in range(num_folds)
+        ]
+        self.flops: Optional[FlopsMetrics] = None
+
+        self.timer = TimeHistory()
+        self.start_date: Optional[str] = None
+        self.end_date: Optional[str] = None
+        self.curr_i_fold: int = 0
+
+        # Config for auto-lifecycle
+        self._label_map = label_map
+        self._save_path = Path(save_path) if save_path else None
+        self._verbose = verbose or display_report
+        self._display_report = display_report
+        self._ended = False
+
+        # Lazy init to avoid circular imports
+        self._display = None
+        self._storage = None
+
+    @property
+    def display(self):
+        if self._display is None:
+            from tracking.display import HistoryDisplay
+            self._display = HistoryDisplay(
+                self,
+                label_map=self._label_map,
+                show_report=self._display_report,
+            )
+        return self._display
+
+    @property
+    def storage(self):
+        if self._storage is None:
+            from tracking.storage import HistoryStorage
+            self._storage = HistoryStorage(self)
+        return self._storage
+
+    # --- Context manager ---
+
+    def __enter__(self):
+        self.start()
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.end()
+        if exc_type is None and self._save_path is not None:
+            self.storage.save(path=self._save_path, label_map=self._label_map)
+
+    # --- Iterator ---
+
+    def __iter__(self):
+        self.start()
+        for fold in self.folds:
+            yield fold
+        self.end()
+
+    # --- Fold management ---
+
+    @property
+    def fold(self) -> FoldHistory:
+        """Shortcut for get_curr_fold()."""
+        return self.folds[self.curr_i_fold]
+
+    def get_curr_fold(self) -> FoldHistory:
+        """Get the current fold."""
+        return self.folds[self.curr_i_fold]
+
+    def step(self):
+        """End current fold and advance to next."""
+        self.folds[self.curr_i_fold].end()
+        if self._verbose:
+            self.display.end_fold()
+        if self.curr_i_fold >= self.num_folds - 1:
+            self.end()
+            return
+        self.curr_i_fold += 1
+        self.folds[self.curr_i_fold].timer.start()
+        if self._verbose:
+            self.display.start_fold()
+
+    # --- Metadata setters ---
+
+    def set_model_parms(self, model_parms: HyperParams):
+        self.model_parms = model_parms
+
+    def set_model_arch(self, model_arch: str):
+        self.model_arch = model_arch
+
+    def set_flops(self, flops: FlopsMetrics):
+        self.flops = flops
+        if self._verbose:
+            self.display.flops()
+
+    # --- Lifecycle ---
+
+    def start(self):
+        self._ended = False
+        self.timer.start()
+        self.start_date = time.strftime("%Y%m%d_%H%M")
+        self.get_curr_fold().start()
+        if self._verbose:
+            self.display.start_fold()
+
+    def end(self):
+        if self._ended:
+            return
+        self._ended = True
+        self.timer.stop()
+        self.end_date = time.strftime("%Y%m%d_%H%M%S")
+        if self._verbose:
+            self.display.end_training()
