+++ src/training/runners/category_cv.py
@@ -12,7 +12,7 @@ from training.helpers import compute_class_weights, setup_optimizer, setup_sched
 from training.runners.category_eval import evaluate
 from training.runners.category_trainer import train
 from common.calc_flops.calculate_model_flops import calculate_model_flops
-from common.ml_history import MLHistory, FlopsMetrics, NeuralParams
+from tracking import MLHistory, FlopsMetrics, NeuralParams
 from common.mps_support import clear_mps_cache
