+++ src/training/runners/category_trainer.py
@@ -9,7 +9,7 @@ from torch.optim.lr_scheduler import LRScheduler
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 
-from common.ml_history.fold import FoldHistory
+from tracking.fold import FoldHistory
 
 
 def train(
