+++ src/training/runners/next_cv.py
@@ -11,8 +11,8 @@ from training.helpers import compute_class_weights, setup_optimizer, setup_sched
 from training.runners.next_eval import evaluate
 from training.runners.next_trainer import train
 from common.calc_flops.calculate_model_flops import calculate_model_flops
-from common.ml_history import MLHistory, FlopsMetrics, NeuralParams
-from common.ml_history.fold import FoldHistory
+from tracking import MLHistory, FlopsMetrics, NeuralParams
+from tracking.fold import FoldHistory
 from common.mps_support import clear_mps_cache
 
 import torch
