+++ src/training/runners/next_trainer.py
@@ -10,7 +10,7 @@ from torch.utils.data import DataLoader
 from tqdm import tqdm
 
 from configs.globals import CATEGORIES_MAP
-from common.ml_history.fold import FoldHistory
+from tracking.fold import FoldHistory
 
 
 def train(
